feat: integrate C3 navigation V2 into Operate behind IS_NAV_V2_ENABLED
Adds the C3 navigation V2 surface (collapsible sidebar + cluster-webapp
breadcrumbs + tool panels) to Operate side-by-side with the existing V1 header,
selected at build time by a new IS_NAV_V2_ENABLED flag (default false). Mirrors
the merged Tasklist pattern (#55050) and the Optimize follow-up (#55562).

Why this shape:
- modules/feature-flags.ts: introduces IS_NAV_V2_ENABLED alongside the existing
  IS_CONVERSATION_HISTORY_ENABLED.
- App/Layout/AppHeader: the former index.tsx moves verbatim to LegacyAppHeader.tsx
  (untouched V1, via git rename) and index.tsx becomes a thin compile-time switch.
- AppHeader/AppHeaderV2 is split into focused files (index + useSidebarChildren +
  useCamundaToolsConfig) rather than one monolith, and memoized. The
  C3UserConfigurationProvider is already mounted at Layout via C3Provider, so the
  V2 header calls useCamundaTools / useClusterWebappBreadcrumbs directly.
- Sidebar mirrors the existing navbar: Dashboard, Processes, Decisions, and an
  Operations group-item (Batch operations / Operations log). Operations has no
  own page, so the parent links to its primary child (Batch operations) and
  stays selected whenever either child is active, matching the Optimize Analysis
  pattern; the V1 small-screen matchMedia flattening is dropped since the sidebar
  group handles it. Active state comes from useCurrentPage().
- Breadcrumbs use useClusterWebappBreadcrumbs({currentApp: 'operate'}); no
  invented modeler-URL helper.
- styled.tsx PageContent offsets via --c3-sidebar-width (fallback 0) so V1 is
  unchanged; the existing --cds-spacing-09 top padding is kept.
- Operate has no i18n framework, so labels stay hardcoded English (existing
  convention) and C3's English sidebar-label defaults apply.
- Tests: the 5 existing AppHeader tests are unchanged (they exercise the V1
  surface through the switch with the flag off); a new headerV2 test covers the
  V2 surface with the flag mocked on.
