+++ operate/client/src/modules/feature-flags.ts
@@ -9,4 +9,6 @@
 const IS_CONVERSATION_HISTORY_ENABLED =
   localStorage.getItem('FEATURE_CONVERSATION_HISTORY_ENABLED') === 'true';
 
-export {IS_CONVERSATION_HISTORY_ENABLED};
+const IS_NAV_V2_ENABLED = false;
+
+export {IS_CONVERSATION_HISTORY_ENABLED, IS_NAV_V2_ENABLED};
